feat + fix: die when touch sprite and fix render sprite order
